meeting-registration/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Limiter==3.5.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
redis==5.0.1
gunicorn==21.2.0
flask-caching
"celery[redis]"